Refactored backend structure and added production deployment files
