+++ driving-day-app-backend/.gitignore
@@ -64,10 +64,11 @@ ehthumbs.db
 .ipynb_checkpoints
 
 .env
-fsae_backend/fsae_backend_app/firebase/serviceAccountKey.json
+fsae_backend_app/firebase/serviceAccountKey.json
 
 # Ignore all .csv files
 *.csv
 
 # Ignore all .ld files
 *.ld
+.vercel
