+++ driving-day-app-backend/.vercelignore
@@ -1,0 +1,18 @@
+# Ignore virtual environment
+venv/
+
+# Ignore database files
+*.sqlite3
+
+# Ignore Python cache
+__pycache__/
+*.pyc
+
+# Ignore static files (if not used)
+staticfiles/
+
+# Ignore local environment files
+.env
+
+# Ignore large data or debug directories
+fsae_backend_app/ld_parser/data/
