--- driving-day-app-backend/fsae_backend/fsae_backend_app/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class FsaeBackendAppConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'fsae_backend_app'
