--- driving-day-app-backend/fsae_backend/fsae_backend_app/tests.py
@@ -1,3 +1,0 @@
-from django.test import TestCase
-
-# Create your tests here.
