+++ driving-day-app-backend/requirements.txt
@@ -1,47 +1,39 @@
 asgiref==3.8.1
-autopep8==2.3.1
-CacheControl==0.14.0
+CacheControl==0.14.1
 cachetools==5.5.0
 certifi==2024.8.30
 cffi==1.17.1
 charset-normalizer==3.4.0
 cryptography==43.0.3
-Django==5.1.2
-django-cors-headers==4.5.0
+Django==5.1.3
+django-cors-headers==4.6.0
 djangorestframework==3.15.2
-firebase-admin==6.5.0
-google-api-core==2.21.0
-google-api-python-client==2.149.0
-google-auth==2.35.0
+firebase-admin==6.6.0
+google-api-core==2.23.0
+google-api-python-client==2.153.0
+google-auth==2.36.0
 google-auth-httplib2==0.2.0
 google-cloud-core==2.4.1
 google-cloud-firestore==2.19.0
 google-cloud-storage==2.18.2
 google-crc32c==1.6.0
 google-resumable-media==2.7.2
-googleapis-common-protos==1.65.0
-grpcio==1.67.0
-grpcio-status==1.67.0
+googleapis-common-protos==1.66.0
+grpcio==1.68.0
+grpcio-status==1.68.0
 httplib2==0.22.0
 idna==3.10
 msgpack==1.1.0
-numpy==2.1.2
-pandas==2.2.3
-proto-plus==1.24.0
-protobuf==5.28.2
+proto-plus==1.25.0
+protobuf==5.28.3
 pyasn1==0.6.1
 pyasn1_modules==0.4.1
-pycodestyle==2.12.1
 pycparser==2.22
-PyJWT==2.9.0
+PyJWT==2.10.0
 pyparsing==3.2.0
-python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
-pytz==2024.2
 requests==2.32.3
 rsa==4.9
-six==1.16.0
-sqlparse==0.5.1
-tzdata==2024.2
+sqlparse==0.5.2
 uritemplate==4.1.1
 urllib3==2.2.3
