+++ driving-day-app-backend/vercel.json
@@ -1,0 +1,15 @@
+{
+    "version": 2,
+    "builds": [
+        {
+            "src": "vercel_python.py",
+            "use": "@vercel/python"
+        }
+    ],
+    "routes": [
+        {
+            "src": "/(.*)",
+            "dest": "/vercel_python.py"
+        }
+    ]
+}
