+++ driving-day-app-backend/vercel_python.py
@@ -1,0 +1,1 @@
+from fsae_backend.asgi import application as app
